fix: transactions issue
