+++ core/src/commonMain/kotlin/com/powersync/db/PowerSyncDatabaseImpl.kt
@@ -226,13 +226,11 @@ internal class PowerSyncDatabaseImpl(
         mapper: (SqlCursor) -> RowType,
     ): Flow<List<RowType>> = internalDb.watch(sql, parameters, mapper)
 
-    @SuspendInterop.Disabled
     override suspend fun <R> readTransaction(callback: suspend (tx: PowerSyncTransaction) -> R): R =
         internalDb.readTransaction { tx ->
             callback(tx)
         }
 
-    @SuspendInterop.Disabled
     override suspend fun <R> writeTransaction(callback: suspend (tx: PowerSyncTransaction) -> R): R =
         internalDb.writeTransaction { tx ->
             callback(tx)
